pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonomous-security-ai"
version = "1.0.0"
description = "Autonomous Security AI: Self-Learning Defense Agent"
readme = "README.md"
authors = [{name = "Security AI Team", email = "security@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "pydantic>=1.8.0",
    "typer>=0.4.0",
    "rich>=10.0.0",
    "joblib>=1.1.0",
    "PyYAML>=6.0",
    "pytest>=6.2.0"
]

[project.optional-dependencies]
dev = ["pytest", "black", "flake8"]

[project.scripts]
security-ai = "src.main:main"

[project.urls]
Homepage = "https://github.com/your-username/autonomous-security-ai"
Repository = "https://github.com/your-username/autonomous-security-ai"